move sceneviews.h out of scene.h
