+++ Source/Engine/VQEngine_Input.cpp
@@ -18,6 +18,7 @@
 #include "VQEngine.h"
 #include "Math.h"
 #include "Scene/Scene.h"
+#include "Scene/SceneViews.h"
 #include "../Scenes/Scenes.h" // scene instances
 
 #include "GPUMarker.h"
