+++ Source/Renderer/Rendering/RenderPass/ObjectIDPass.cpp
@@ -19,6 +19,7 @@
 #include "ObjectIDPass.h"
 #include "Renderer/Renderer.h"
 #include "Engine/Scene/Scene.h"
+#include "Engine/Scene/SceneViews.h"
 #include "Shaders/LightingConstantBufferData.h"
 #include "Libs/VQUtils/Source/utils.h"
 #include "Engine/GPUMarker.h"
