+++ Source/Scenes/TerrainScene.cpp
@@ -1,8 +1,9 @@
 #include "Scenes.h"
 
-#include "../Engine/Core/Input.h"
+#include "Engine/Scene/SceneViews.h"
+#include "Engine/Core/Input.h"
 
-#include "../Libs/VQUtils/Source/utils.h"
+#include "Libs/VQUtils/Source/utils.h"
 
 using namespace DirectX;
